Rename gvm_connection python module to gmp
In future this module will be splitted into a gmp part and a connection
part. But currently it contains the classes passed to our scripts as the
gmp global object.
