+++ gmp/clients/gvm_dialog.py
@@ -26,9 +26,9 @@ import sys
 from dialog import Dialog
 from lxml import etree
 from gmp.helper import get_version
-from gmp.gvm_connection import (SSHConnection,
-                                TLSConnection,
-                                UnixSocketConnection)
+from gmp.gmp import (SSHConnection,
+                     TLSConnection,
+                     UnixSocketConnection)
 
 __version__ = get_version()
