+++ gmp/clients/gvm_pyshell.py
@@ -31,9 +31,9 @@ import os
 import sys
 from gmp.helper import get_version
 from lxml import etree
-from gmp.gvm_connection import (SSHConnection,
-                                TLSConnection,
-                                UnixSocketConnection)
+from gmp.gmp import (SSHConnection,
+                     TLSConnection,
+                     UnixSocketConnection)
 
 
 __version__ = get_version()
